+++ src/GlucoPilot.Api/Endpoints/Ingredients/NewIngredient/NewIngredientResponse.cs
@@ -8,10 +8,10 @@ public sealed record NewIngredientResponse
     public required Guid Id { get; set; }
     public required DateTimeOffset Created { get; set; }
     public required string Name { get; set; }
-    public int Carbs { get; set; }
-    public int Protein { get; set; }
-    public int Fat { get; set; }
-    public int Calories { get; set; }
+    public decimal Carbs { get; set; }
+    public decimal Protein { get; set; }
+    public decimal Fat { get; set; }
+    public decimal Calories { get; set; }
     public required UnitOfMeasurement Uom { get; set; }
     public DateTimeOffset? Updated { get; set; }
 }
